+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Python: Added PFCOUNT command ([#1493](https://github.com/aws/glide-for-redis/pull/1493))
 * Python: Added PFMERGE command ([#1497](https://github.com/aws/glide-for-redis/pull/1497))
 * Node: Added SINTER command ([#1500](https://github.com/aws/glide-for-redis/pull/1500))
+* Python: Added LASTSAVE command ([#1509](https://github.com/aws/glide-for-redis/pull/1509))
 
 ## 0.4.0 (2024-05-26)